+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 # Unreleased
+- refactor DuckDB::ScalarFunction to use vector_set_value_at helper.
 - fix DuckDB::ScalarFunction NULL input handling.
 - fix DuckDB::ScalarFunction INTEGER output type (int32 vs int64).
 - add DuckDB::ScalarFunction#add_parameter.